Add new project Task1-02
